将所有数字类型的单元格内容解析为Long行，将导致浮点数丢失小数部分

+++ src/main/java/cn/xuemengzihe/util/excel/ParseExcelFile.java
@@ -76,7 +76,7 @@ public class ParseExcelFile implements Closeable {
 			value = cell.getCellFormula();
 			break;
 		case Cell.CELL_TYPE_NUMERIC:
-			value = cell.getNumericCellValue() + "";
+			value = Double.valueOf(cell.getNumericCellValue()).longValue() + "";
 			break;
 		case Cell.CELL_TYPE_STRING:
 			value = cell.getStringCellValue();
